Moved detail type check into util function and created Detail type
